move to prompts dictionary
